add forms in index and refactor
