+++ src/shared/ui/forms/FormCheckbox.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { FieldValues, UseControllerProps, useController } from 'react-hook-form';
 
 import { Checkbox, CheckboxPropsProps } from '@/shared/ui';
@@ -5,7 +6,7 @@ import { Checkbox, CheckboxPropsProps } from '@/shared/ui';
 type Props<T extends FieldValues> = Omit<CheckboxPropsProps, 'checked' | 'name' | 'onBlur' | 'onCheckedChange'> &
   UseControllerProps<T>;
 
-function FormCheckbox<T extends FieldValues>({ control, name, ...rest }: Props<T>) {
+export const FormCheckbox = <T extends FieldValues>({ control, name, ...rest }: Props<T>) => {
   const {
     field: { onChange, value = rest.defaultChecked, ...field }
   } = useController({
@@ -14,6 +15,4 @@ function FormCheckbox<T extends FieldValues>({ control, name, ...rest }: Props<T
   });
 
   return <Checkbox {...rest} {...field} checked={value} onCheckedChange={onChange} />;
-}
-
-export default FormCheckbox;
+};
