+++ src/shared/ui/forms/FormTextfield.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { FieldValues, UseControllerProps, useController } from 'react-hook-form';
 
 import { Input, InputProps } from '@/shared/ui';
