+++ src/shared/ui/index.ts
@@ -3,6 +3,8 @@ export * from './button/Button';
 export * from './card/Card';
 export * from './checkbox/Checkbox';
 export * from './datePicker/ui/DatePicker';
+export * from './forms/FormCheckbox';
+export * from './forms/FormTextfield';
 export * from './input/Input';
 export * from './mobileMenu/MobileMenu';
 export * from './modal/Modal';
